requirements.txt
argcomplete==1.9.3
boto3==1.9.130
botocore==1.12.130
certifi==2019.3.9
cfn-flip==1.2.0
chardet==3.0.4
Click==7.0
docutils==0.14
durationpy==0.5
Flask==1.0.2
Flask-Cors==3.0.7
future==0.16.0
hjson==3.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
MarkupSafe==1.1.1
numpy==1.16.2
pandas==0.24.2
placebo==0.9.0
pymongo==3.7.2
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.1
PyYAML==5.1
requests==2.21.0
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.6
Unidecode==1.0.23
urllib3==1.24.1
Werkzeug==0.15.2
wsgi-request-logger==0.4.6
zappa==0.48.2
